requirements.txt
aiofiles==24.1.0
aiohttp==3.8.4
asyncpg==0.29.0
discord.py==2.2.2
loguru==0.6.0
